refactor 1. amount, pay 관련 자료형 int 에서 Long으로 변경 2. Session 에 session 찾는 메서드 추가

+++ src/main/java/nextstep/courses/domain/Provide.java
@@ -25,7 +25,7 @@ public class Provide {
         }
     }
     
-    public void applyPaid(int enrolledCount, int pay) throws CanNotJoinException {
+    public void applyPaid(int enrolledCount, Long pay) throws CanNotJoinException {
         if(this.type == ProvideType.FREE) {
             throw new CanNotJoinException("유료 강의는 결제를 해야한다");
         }
